# api/openapi.yaml
# API contract baseline (D6, AGENTS.md §7). Generated by springdoc (GET /v3/api-docs.yaml) and
# committed. ApiContractIT fails ./gradlew check on any breaking change against this file.
# Changes are additive only; update this file in the same commit as the API change.
openapi: 3.1.0
info:
  title: URL shortener
  version: "1"
servers:
- url: /
paths:
  /api/keys:
    post:
      operationId: createApiKey
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedApiKey"
          description: Created
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
      security:
      - apiKey: []
      summary: Issue an owner API key (admin only)
      tags:
      - keys
  /api/links:
    post:
      operationId: createLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedLink"
          description: Created
          headers:
            Location:
              schema:
                type: string
              style: simple
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Creation limit reached
          headers:
            Retry-After:
              schema:
                type: integer
              style: simple
        "503":
          $ref: "#/components/responses/Problem"
      security:
      - apiKey: []
      summary: Create a short link (owner only)
      tags:
      - links
  /api/links/{code}:
    delete:
      operationId: deleteLink
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
      security:
      - apiKey: []
      summary: Delete one of the caller's links
      tags:
      - links
    get:
      operationId: getLink
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
          description: OK
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
      security:
      - apiKey: []
      summary: Read one of the caller's links
      tags:
      - links
  /{code}:
    get:
      operationId: redirect
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "302":
          description: Found
          headers:
            Location:
              schema:
                type: string
                format: uri
              style: simple
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
      summary: Follow a short link (public)
      tags:
      - redirect
components:
  responses:
    Problem:
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
      description: Error (RFC 9457 problem details)
  schemas:
    CreateLinkRequest:
      type: object
      properties:
        targetUrl:
          type: string
          maxLength: 2048
      required:
      - targetUrl
    CreatedApiKey:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        key:
          type: string
      required:
      - createdAt
      - id
      - key
    CreatedLink:
      type: object
      properties:
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        shortUrl:
          type: string
          format: uri
        targetUrl:
          type: string
          format: uri
      required:
      - code
      - createdAt
      - shortUrl
      - targetUrl
    Link:
      type: object
      properties:
        code:
          type: string
        createdAt:
          type: string
          format: date-time
        shortUrl:
          type: string
          format: uri
        status:
          type: string
          description: Clients must tolerate values not listed here
          x-extensible-enum:
          - ACTIVE
        targetUrl:
          type: string
          format: uri
      required:
      - code
      - createdAt
      - shortUrl
      - status
      - targetUrl
    Problem:
      type: object
      properties:
        code:
          type: string
          description: Stable error code
        instance:
          type: string
        reason:
          type: string
          description: url-rejected only
        requestId:
          type: string
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          description: urn:url-shortener:problem:<code>
      required:
      - code
      - requestId
      - status
      - title
      - type
  securitySchemes:
    apiKey:
      description: "API key: Authorization: Bearer usk_<prefix>_<secret>"
      scheme: bearer
      type: http